requirements.txt
tensorflow==2.20.0
numpy
pillow
flask
flask-cors
matplotlib
scikit-learn
gunicorn
